requirements.txt
asgiref==3.3.0
Django==3.1.3
django-filter==2.4.0
django-rest==0.8.6
djangorestframework==3.12.2
Markdown==3.3.3
Pillow==8.0.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
